rename adoptions-list file to my-pets-list
