+++ apps/mobile/__tests__/screens/adoptions/my-pets-list.test.tsx
@@ -1,4 +1,4 @@
-import { MyPetsList } from '@/screens/adoptions/compose/adoptions-list';
+import { MyPetsList } from '@/screens/adoptions/compose/my-pets-list';
 import { renderWithProviders } from '@/test/test-utils';
 
 describe('<MyPetsList />', () => {
